pyproject.toml
[project]
name = "objdet"
version = "0.1.0"
description = "Production-grade object detection training framework built on PyTorch Lightning"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
keywords = [
    "object-detection",
    "pytorch",
    "pytorch-lightning",
    "deep-learning",
    "computer-vision",
    "yolo",
    "faster-rcnn",
    "mlops",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]

dependencies = [
    # Core ML
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "lightning>=2.2.0",
    "torchmetrics>=1.3.0",

    # YOLO
    "ultralytics>=8.1.0",

    # Data
    "litdata>=0.2.0",
    "albumentations>=1.4.0",
    "sahi>=0.11.0",

    # Experiment Tracking
    "mlflow>=2.10.0",
    "tensorboard>=2.16.0",

    # Optimization & Inference
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "safetensors>=0.4.0",

    # Serving
    "litserve>=0.2.0",
    "pydantic>=2.6.0",

    # Job Queue
    "celery>=5.3.0",

    # Hyperparameter Tuning
    "optuna>=3.5.0",

    # Logging
    "loguru>=0.7.0",
    "rich>=13.7.0",

    # Configuration
    "jsonargparse[signatures]>=4.27.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # Modern HTTP client (replaces requests)
    "httpx>=0.27.0",

    # Modern datetime library (replaces stdlib datetime)
    "whenever>=0.8.0",

    # Utilities
    "numpy>=1.26.0",
    "pillow>=10.2.0",
    "tqdm>=4.66.0",
    "ensemble-boxes>=1.0.9",
]

[project.optional-dependencies]
tensorrt = ["tensorrt>=8.6.0"]
thunder = ["lightning-thunder>=0.2.0"]
all = ["objdet[tensorrt,thunder]"]

[dependency-groups]
dev = [
    { include-group = "test" },
    { include-group = "lint" },
    { include-group = "docs" },
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
lint = ["pre-commit>=3.6.0", "ruff>=0.3.0"]
docs = [
    "sphinx>=7.2.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-rtd-theme>=2.0.0",
    "myst-parser>=2.0.0",
]

[project.scripts]
objdet = "objdet.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/objdet"
Documentation = "https://objdet.readthedocs.io"
Repository = "https://github.com/yourusername/objdet"
Issues = "https://github.com/yourusername/objdet/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/objdet"]

# ============================================================================
# Ruff Configuration
# ============================================================================
[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate (commented-out code)
    "PL",  # Pylint
    "RUF", # Ruff-specific rules
    "D",   # pydocstyle
    "ANN", # flake8-annotations
]
ignore = [
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "D107",    # Missing docstring in __init__
    "ANN101",  # Missing type annotation for self (removed in newer ruff)
    "ANN102",  # Missing type annotation for cls (removed in newer ruff)
    "ANN401",  # Any type in *args/**kwargs - often unavoidable
    "PLR0913", # Too many arguments in function definition
    "PLC0415", # Import not at top-level - intentional for lazy loading
    "PLR2004", # Magic value comparison - too noisy
    "PLW0602", # Global without assignment - false positive for guards
    "ARG001",  # Unused function argument - often required by interface
    "ARG002",  # Unused method argument - often required by interface
    "ARG005",  # Unused lambda argument - often required for compatibility
    "TC001",   # Type checking block suggestions - too strict
    "TC002",   # Type checking block suggestions - too strict
    "TC003",   # Type checking block suggestions - too strict
    "TC004",   # Move import out of type-checking block - style preference
    "UP040",   # TypeAlias annotation instead of type keyword - style preference
    "UP046",   # Generic[T] vs type parameters - style preference
    "RUF005",  # List concatenation - explicit sometimes clearer
    "RUF012",  # Mutable class attributes - ClassVar not always needed
    "RUF015",  # next(iter()) suggestion - list()[0] is fine
    "RUF022",  # __all__ not sorted - grouped by category intentionally
    "RUF059",  # Unused unpacked variable - often needed for API
    "RUF100",  # Unused noqa - false positives
    "SIM108",  # Ternary operator suggestion - explicit if/else often clearer
    "PTH123",  # Path.open() suggestion - open() is fine
    "B007",    # Unused loop control variable - often needed
    "B904",    # raise from - not always needed for pruned exceptions
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "ANN", "PLR2004"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

# ============================================================================
# ty Configuration (Type Checking)
# ============================================================================
[tool.ty.environment]
python-version = "3.12"

[tool.ty.rules]
possibly-unresolved-reference = "warn"

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers", "--strict-config"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests requiring GPU",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/objdet"]
branch = true
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true

# ============================================================================
# Git-cliff Configuration (Changelog)
# ============================================================================
# Note: Full git-cliff config is in cliff.toml
